scheduling/requirements.txt
apache-airflow>=2.7.0 